разбил js на модули
